// web-angular-nodejs/src/app/core/models/dashboardDetails.model.ts
export class DashbordDetailsModel {
  examDate: string = '';
  date: string = '';
  total: number = 0;
  totalRegistration: number = 0;
  active: number = 0;
  inactive: number = 0;
  adminLogout: number = 0;
  back: number = 0;
  completed: number = 0;
  appeared: number = 0;
  notAppeared: number = 0;
  focussedOut: number = 0;
}